requirements.txt
requests>=2.28.1
python-whois>=0.8.0
tldextract>=3.4.0
validators>=0.20.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.9
idna>=3.4
urllib3>=1.26.15
colorama>=0.4.6
click>=8.1.3
pytest>=7.3.1 